requirements.txt
torch
torchvision
Streamlit
Pillow
matplotlib
numpy
fastapi
uvicorn